feat: agregar metodo getProfile
